# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Take Home Test API",
      "description": "Documentation for Take Home Test API, Bambang Rahmat (NodeJs)",
      "version": "1.0.0"
    },
...
    "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "securityDefinitions": {
      "bearerAuth": {
          "type": "apiKey",
          "in": "header",
          "name": "Authorization"
      }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
    "paths": {
...
      "/profile": {
        "get": {
          "tags": ["1. Membership"],
          "summary": "Get user profile",
          "description": "API Profile Private (memerlukan Token untuk mengaksesnya). Digunakan untuk mendapatkan informasi profil User.\n\nKetentuan:\n- Service ini harus menggunakan Bearer Token JWT untuk mengaksesnya.\n- Tidak ada parameter email di query param URL ataupun request body, parameter email diambil dari payload JWT yang didapatkan dari hasil login.\n- Handling Response sesuai dokumentasi Response di bawah.",
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Request Successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 0
                      },
                      "message": {
                        "type": "string",
                        "example": "Sukses"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "email": {
                            "type": "string",
                            "example": "user@testpraktek.com"
                          },
                          "first_name": {
                            "type": "string",
                            "example": "User"
                          },
                          "last_name": {
                            "type": "string",
                            "example": "Bambang"
                          },
                          "profile_image": {
                            "type": "string",
                            "format": "uri",
                            "example": "https://yoururlapi.com/profile.jpeg"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized - Token tidak valid atau kadaluwarsa",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 108
                      },
                      "message": {
                        "type": "string",
                        "example": "Token tidak valid atau kadaluwarsa"
                      },
                      "data": {
                        "type": "null",
                        "example": null
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/profile/update": {
        "put": {
          "tags": ["1. Membership"],
          "summary": "Update user profile",
          "description": "Digunakan untuk memperbarui profil user.\n\nKetentuan:\n- Service ini harus menggunakan Bearer Token JWT untuk mengaksesnya.\n- Tidak ada parameter email di query param URL ataupun request body, parameter email diambil dari payload JWT.\n- Handling Response sesuai dokumentasi Response di bawah.",
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "first_name": {
                      "type": "string",
                      "example": "User"
                    },
                    "last_name": {
                      "type": "string",
                      "example": "Nutech"
                    }
                  },
                  "required": ["first_name", "last_name"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Profile updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 0
                      },
                      "message": {
                        "type": "string",
                        "example": "Profile updated successfully"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request - Data tidak valid",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 102
                      },
                      "message": {
                        "type": "string",
                        "example": "Data tidak valid"
                      },
                      "data": {
                        "type": "null",
                        "example": null
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized - Token tidak valid atau kadaluwarsa",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 108
                      },
                      "message": {
                        "type": "string",
                        "example": "Token tidak valid atau kadaluwarsa"
                      },
                      "data": {
                        "type": "null",
                        "example": null
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/profile/image": {
        "put": {
          "tags": ["1. Membership"],
          "summary": "Update user profile image",
          "description": "Digunakan untuk memperbarui foto profil user.\n\nKetentuan:\n- Service ini harus menggunakan Bearer Token JWT untuk mengaksesnya.\n- Format file yang diperbolehkan: JPG, PNG.\n- Ukuran maksimal file: 2MB.",
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile_image": {
                      "type": "string",
                      "format": "binary",
                      "description": "File gambar untuk foto profil"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Profile image updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 0
                      },
                      "message": {
                        "type": "string",
                        "example": "Profile image updated successfully"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request - Format file tidak valid atau ukuran terlalu besar",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 102
                      },
                      "message": {
                        "type": "string",
                        "example": "Format file tidak valid atau ukuran terlalu besar"
                      },
                      "data": {
                        "type": "null",
                        "example": null
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized - Token tidak valid atau kadaluwarsa",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 108
                      },
                      "message": {
                        "type": "string",
                        "example": "Token tidak valid atau kadaluwarsa"
                      },
                      "data": {
                        "type": "null",
                        "example": null
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }